use static query to fetch data for banner
