+++ content/banner/slides.yml
@@ -3,4 +3,4 @@ slides:
   - text: "Don't miss the session by Lukasz Gornicki at ASC on October 17 at 14:00PST"
     url: "https://asc2019.sched.com/event/TUmg/specifications-in-the-core-of-the-extensibility-lukasz-gornicki-sap"
     startDate: "13/10/2019"
-    endDate: "17/10/2019"
+    endDate: "22/10/2019"
