+++ package.json
@@ -46,9 +46,9 @@
     "@fortawesome/free-regular-svg-icons": "^5.9.0",
     "@fortawesome/free-solid-svg-icons": "^5.9.0",
     "@fortawesome/react-fontawesome": "^0.1.4",
-    "@kyma-project/documentation-component": "0.3.0",
     "@kyma-project/dc-markdown-render-engine": "0.3.0",
     "@kyma-project/dc-open-api-render-engine": "0.3.0",
+    "@kyma-project/documentation-component": "0.3.0",
     "@octokit/rest": "^16.25.1",
     "@rooks/use-timeout": "^3.3.0",
     "babel-plugin-styled-components": "^1.10.0",
@@ -61,14 +61,15 @@
     "gatsby-plugin-netlify": "^2.0.16",
     "gatsby-plugin-netlify-cache": "^1.1.0",
     "gatsby-plugin-react-helmet": "^3.0.2",
-    "gatsby-plugin-sitemap": "^2.1.0",
     "gatsby-plugin-robots-txt": "^1.5.0",
+    "gatsby-plugin-sitemap": "^2.1.0",
     "gatsby-plugin-styled-components": "^3.0.4",
     "gatsby-plugin-typescript": "^2.0.3",
     "gatsby-plugin-typography": "^2.2.5",
     "gatsby-remark-embed-video": "^1.7.0",
     "gatsby-source-filesystem": "^2.1.5",
     "gatsby-transformer-remark": "^2.2.5",
+    "gatsby-transformer-yaml-full": "^0.2.0",
     "lodash.throttle": "^4.1.1",
     "netlify-lambda": "^1.4.5",
     "node-fetch": "^2.5.0",
